tiffcrop subroutines require a larger buffer -- change buffersize to size_t
Buffer overflows have been fixed by making the buffers at least 3 bytes larger than the image itself needs.
This is required for some tiffcrop conversion subroutines.

In addition, some integer casting warnings have been fixed.

Closes issues #271, #381, #386, #388, #389 and #435.

+++ tools/tiffcrop.c
@@ -237,7 +237,7 @@ struct offset {
  */
 
 struct  buffinfo {
-  uint64_t size;           /* size of this buffer */
+  size_t size;           /* size of this buffer */
   unsigned char *buffer; /* address of the allocated buffer */
 };
